% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Load Zenodo files into a running session}
\usage{
download(name, format = "rds", url)
}
\arguments{
\item{name}{A character value with the name of the data set to be downloaded.}

\item{format}{Format of data. Altenatives.}

\item{url}{A character value with the link to the source. Only used if
missing name of data set.}
}
\description{
Download data stored in Zenodo into a running R session.

List alternative formats.

Refer to a new function listing data sets.
}
